enable TaskSendNodePublish

+++ Common_CSAT/Inc/TaskSendNodePortList.hpp
@@ -4,6 +4,8 @@
 #include "Task.hpp"
 #include "RegistrationManager.hpp"
 
+#include <cstddef>
+
 #include "nunavut_assert.h"
 #include "uavcan/node/port/List_1_0.h"
 
@@ -34,12 +36,12 @@ void TaskSendNodePortList<Adapters...>::handleTaskImpl()
     uavcan_node_port_List_1_0 data;
     size_t publication_size = registration_manager_->getPublications().size();
     data.publishers.sparse_list.count = publication_size;
-    for(uint i=0; i<publication_size; ++i) data.publishers.sparse_list.elements[i].value = registration_manager_->getPublications()[i];
+    for(uint16_t i=0; i<publication_size; ++i) data.publishers.sparse_list.elements[i].value = registration_manager_->getPublications()[i];
     data.publishers._tag_ = uavcan_node_port_SubjectIDList_1_0_sparse_list_ARRAY_IS_VARIABLE_LENGTH_;
     
     size_t subscription_size = registration_manager_->getSubscriptions().size();
     data.subscribers.sparse_list.count = subscription_size;
-    for(uint i=0; i<subscription_size; ++i) data.subscribers.sparse_list.elements[i].value = registration_manager_->getSubscriptions()[i];
+    for(uint16_t i=0; i<subscription_size; ++i) data.subscribers.sparse_list.elements[i].value = registration_manager_->getSubscriptions()[i];
     data.subscribers._tag_ = uavcan_node_port_SubjectIDList_1_0_sparse_list_ARRAY_IS_VARIABLE_LENGTH_;
